Rearrange WCC and introduce inspectors for WCCValue
