+++ arangod/Pregel/Algos/WCC/WCC.cpp
